file_handling: added main thread restrictions to tests for updating report properties in ZippedHtmlImportProcessSpec
also:
- added tests/implementation for setting the report properties from the json descriptor
- changed report lat/lon properties to NSNumber * to allow for nil

+++ DICE/ReportMapAnnotation.m
@@ -28,8 +28,8 @@
         self.report = report;
         
         CLLocationCoordinate2D coordinate2;
-        coordinate2.latitude = report.lat;
-        coordinate2.longitude = report.lon;
+        coordinate2.latitude = report.lat.doubleValue;
+        coordinate2.longitude = report.lon.doubleValue;
         self.coordinate = coordinate2;
     }
